Feature: Move function from docker api
