+++ api/git/git_service.go
@@ -17,6 +17,7 @@ package git
 
 import (
 	"errors"
+	"fmt"
 	"github.com/marcellodesales/cloner/config"
 	"github.com/marcellodesales/cloner/util"
 	"os"
@@ -76,6 +77,26 @@ func (service GitServiceType) GetRepoLocalPath(gitRepoClone *GitRepoClone, confi
  */
 func (service GitServiceType) GetOrgLocalPath(gitRepoClone *GitRepoClone, config *config.Configuration) string {
 	return path.Join(config.Git.CloneBaseDir, gitRepoClone.Type.GetUserDir())
+}
+
+func (service GitServiceType) VerifyCloneDir(gitRepoClone *GitRepoClone, forceClone bool, config *config.Configuration) error {
+	// The location is provided by the api
+	gitRepoClone.CloneLocation = service.GetRepoLocalPath(gitRepoClone, config)
+	log.Debugf("Verifying if the clone path '%s' exists or exists and is empty", gitRepoClone.CloneLocation)
+
+	if util.DirExists(gitRepoClone.CloneLocation) {
+		if forceClone {
+			util.DeleteDir(gitRepoClone.CloneLocation)
+			return nil
+		}
+
+		// Verify if the user repo is not empty
+		dirIsEmpty, _ := util.IsDirEmpty(gitRepoClone.CloneLocation)
+		if !dirIsEmpty {
+			return errors.New(fmt.Sprintf("clone location '%s' exists and it's not empty", gitRepoClone.CloneLocation))
+		}
+	}
+	return nil
 }
 
 /**
