+++ api/git/git_service_docker.go
@@ -16,33 +16,10 @@ limitations under the License.
 package git
 
 import (
-	"errors"
-	"fmt"
 	"github.com/marcellodesales/cloner/config"
 	"github.com/marcellodesales/cloner/util"
-	log "github.com/sirupsen/logrus"
 	"os"
 )
-
-func (service GitServiceType) VerifyCloneDir(gitRepoClone *GitRepoClone, forceClone bool, config *config.Configuration) error {
-	// The location is provided by the api
-	gitRepoClone.CloneLocation = service.GetRepoLocalPath(gitRepoClone, config)
-	log.Debugf("Verifying if the clone path '%s' exists or exists and is empty", gitRepoClone.CloneLocation)
-
-	if util.DirExists(gitRepoClone.CloneLocation) {
-		if forceClone {
-			util.DeleteDir(gitRepoClone.CloneLocation)
-			return nil
-		}
-
-		// Verify if the user repo is not empty
-		dirIsEmpty, _ := util.IsDirEmpty(gitRepoClone.CloneLocation)
-		if !dirIsEmpty {
-			return errors.New(fmt.Sprintf("clone location '%s' exists and it's not empty", gitRepoClone.CloneLocation))
-		}
-	}
-	return nil
-}
 
 /**
  * Regular git clone execution using docker
